examples/second-brain/pyproject.toml
[project]
name = "second-brain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyyaml",
    "returns",
]
